LAPR22G96-263 #comment US313 implementation updated (not finished yet) #time 40m

+++ src/main/java/lapr/project/data/DataBaseScripts/AverageCargoByYearScript.java
@@ -82,7 +82,7 @@ public class AverageCargoByYearScript {
         }
     }
 
-    public int getNumberOfContainersPerTrip(int mmsi, int date) throws NoContainersInsideThatTripException, NoCargoManifestInThatDateException, NoCargoManifestsWereFoundInThatTripException {
+    public double getNumberOfContainersPerTrip(int mmsi, int date) throws NoContainersInsideThatTripException, NoCargoManifestInThatDateException, NoCargoManifestsWereFoundInThatTripException {
 
         int j = 0;
